Refactor codebase, change folder structure
