+++ .eslintrc.js
@@ -1,3 +1,19 @@
 module.exports = {
-  extends: ['airbnb-base', 'plugin:prettier/recommended']
+  root: true,
+  env: {
+    node: true
+  },
+  extends: [
+    'airbnb-base',
+    'plugin:prettier/recommended',
+    'plugin:vue/essential',
+    '@vue/prettier'
+  ],
+  rules: {
+    'no-console': process.env.NODE_ENV === 'production' ? 'error' : 'off',
+    'no-debugger': process.env.NODE_ENV === 'production' ? 'error' : 'off'
+  },
+  parserOptions: {
+    parser: 'babel-eslint'
+  }
 };
