--- dashboard/.gitignore
@@ -1,24 +1,0 @@
-.DS_Store
-node_modules
-/dist
-/data/commits_count.json
-/data/commits.json
-/data/dashboard.json
-
-# local env files
-.env.local
-.env.*.local
-
-# Log files
-npm-debug.log*
-yarn-debug.log*
-yarn-error.log*
-
-# Editor directories and files
-.idea
-.vscode
-*.suo
-*.ntvs*
-*.njsproj
-*.sln
-*.sw*
