--- dashboard/package.json
@@ -1,38 +1,0 @@
-{
-  "name": "ui",
-  "version": "0.1.0",
-  "private": true,
-  "scripts": {
-    "serve": "vue-cli-service serve",
-    "build": "vue-cli-service build",
-    "lint": "vue-cli-service lint"
-  },
-  "dependencies": {
-    "@vue/ui": "^0.5.2",
-    "chart.js": "^2.7.2",
-    "moment": "^2.22.2",
-    "vue": "^2.5.16",
-    "vue-chartkick": "^0.4.0"
-  },
-  "devDependencies": {
-    "@vue/cli-plugin-babel": "^3.0.0-rc.8",
-    "@vue/cli-plugin-eslint": "^3.0.0-rc.8",
-    "@vue/cli-service": "^3.0.0-rc.8",
-    "@vue/eslint-config-prettier": "^3.0.0-rc.8",
-    "lint-staged": "^7.2.0",
-    "vue-template-compiler": "^2.5.16"
-  },
-  "gitHooks": {
-    "pre-commit": "lint-staged"
-  },
-  "lint-staged": {
-    "*.js": [
-      "vue-cli-service lint",
-      "git add"
-    ],
-    "*.vue": [
-      "vue-cli-service lint",
-      "git add"
-    ]
-  }
-}
