+++ package.json
@@ -2,28 +2,40 @@
   "name": "git-scene",
   "version": "1.0.6",
   "description": "Code contribution graph for every git repo.",
-  "main": "src/index.js",
-  "scripts": {
-    "start": "./bin/git-scene.js",
-    "ui": "cd dashboard && npm run serve",
-    "lint": "eslint src --fix",
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
+  "main": "cli/index.js",
   "bin": {
     "git-scene": "./bin/git-scene.js"
   },
   "author": "Prashant Andani",
   "license": "MIT",
+  "scripts": {
+    "start": "./bin/git-scene.js",
+    "ui": "npm run serve",
+    "lint": "vue-cli-service lint",
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "serve": "vue-cli-service serve",
+    "build": "vue-cli-service build"
+  },
   "dependencies": {
+    "@vue/ui": "^0.5.2",
+    "chart.js": "^2.7.2",
+    "moment": "^2.22.2",
+    "vue": "^2.5.16",
+    "vue-chartkick": "^0.4.0",
     "cli-gh-cal": "^1.4.9",
     "colors": "^1.3.2",
     "express": "^4.16.3",
     "http": "0.0.0",
-    "moment": "^2.22.2",
     "nodegit": "^0.22.2",
     "promise": "^8.0.1"
   },
   "devDependencies": {
+    "@vue/cli-plugin-babel": "^3.0.0-rc.8",
+    "@vue/cli-plugin-eslint": "^3.0.0-rc.8",
+    "@vue/cli-service": "^3.0.0-rc.8",
+    "@vue/eslint-config-prettier": "^3.0.0-rc.8",
+    "lint-staged": "^7.2.0",
+    "vue-template-compiler": "^2.5.16",
     "babel-preset-env": "^1.7.0",
     "eslint": "^5.5.0",
     "eslint-config-airbnb-base": "^13.1.0",
@@ -32,5 +44,18 @@
     "eslint-plugin-prettier": "^2.6.2",
     "path": "^0.12.7",
     "prettier": "^1.14.2"
+  },
+  "gitHooks": {
+    "pre-commit": "lint-staged"
+  },
+  "lint-staged": {
+    "*.js": [
+      "vue-cli-service lint",
+      "git add"
+    ],
+    "*.vue": [
+      "vue-cli-service lint",
+      "git add"
+    ]
   }
 }
